User can delete a watchlist (#4)
* Add Shadcn-svelte component (dialog)

* You can delete watchlists by
1) Selecting the watchlist you want to delete
2) Using the "Delete" button and following dialog prompts

Deleting a watchlist will reset the currently selected watchlist to the first one (sorted by name).

One thing to note is that since there are more than one thing affected the selected watchlist name, the WatchlistCommandSection is no longer bindable w.r.t the selected watchlist: it is fully controlled by it's consumer passing in the selected watchlist via props.

* Refactor the watchlist.ts filein api to use folder for watchlists, for easier re-use and readability

A lot of the diff is just updating imports